Refactor: move style imports to separate file
