+++ client/components/add-product.js
@@ -24,29 +24,11 @@ import {
   TableRow,
 } from '@material-ui/core'
 
-const styles = (theme) => ({
-  root: {
-    '& > *': {
-      margin: theme.spacing(1),
-      width: '50ch',
-      display: 'flex',
-      flexWrap: 'wrap',
-    },
-  },
-  demo: {
-    backgroundColor: theme.palette.background.paper,
-  },
-  title: {
-    margin: theme.spacing(4, 0, 2),
-  },
-  itemText: {
-    fontSize: '.9em',
-    fontWeight: 100,
-  },
-})
+//Material UI styling
+import styles from '../admin-styles'
 
 const tableColumns = [
-  {id: 'id', label: 'ID', minWidth: 60},
+  {id: 'id', label: 'ID55', minWidth: 60},
   {id: 'name', label: 'NAME', minWidth: 100},
   {id: 'style', label: 'STYLE', minWidth: 100},
   {id: 'price', label: 'PRICE', minWidth: 60},
