+++ client/components/all-users.js
@@ -17,26 +17,29 @@ import {
   TableRow,
 } from '@material-ui/core'
 
-const styles = (theme) => ({
-  root: {
-    '& > *': {
-      margin: theme.spacing(1),
-      width: '50ch',
-      display: 'flex',
-      flexWrap: 'wrap',
-    },
-  },
-  demo: {
-    backgroundColor: theme.palette.background.paper,
-  },
-  title: {
-    margin: theme.spacing(4, 0, 2),
-  },
-  itemText: {
-    fontSize: '.9em',
-    fontWeight: 100,
-  },
-})
+//Material UI styling
+import styles from '../admin-styles'
+
+// const styles = (theme) => ({
+//   root: {
+//     '& > *': {
+//       margin: theme.spacing(1),
+//       width: '50ch',
+//       display: 'flex',
+//       flexWrap: 'wrap',
+//     },
+//   },
+//   demo: {
+//     backgroundColor: theme.palette.background.paper,
+//   },
+//   title: {
+//     margin: theme.spacing(4, 0, 2),
+//   },
+//   itemText: {
+//     fontSize: '.9em',
+//     fontWeight: 100,
+//   },
+// })
 
 const tableColumns = [
   {id: 'id', label: 'ID', minWidth: 60},
